+++ knip.json
@@ -6,7 +6,8 @@
         "packages/pwa/src/index.tsx",
         "packages/extension/src/index.tsx",
         "packages/extension/src/background.ts",
-        "packages/functions/src/index.ts"
+        "packages/functions/src/index.ts",
+        "packages/rssServer/src/server.ts"
       ],
       "project": "packages/{pwa,extension,functions,rssServer,shared,sharedClient,sharedServer}/src/**/*.{ts,tsx,js,jsx}",
       "ignoreDependencies": [
